Revert "Arm backend: Support per-channel in TOSA.RESCALE (#15267)"
This reverts commit f7ca57e3be96d1378107892b2017adc445dd56cf.

+++ backends/arm/test/misc/test_tosa_dialect_dw_conv2d.py
@@ -32,7 +32,7 @@ def test_depthwise_conv2d_tosa_INT():
                 8,
             ),
             (1, 16, 20, 20),
-            torch.int32,
+            torch.int8,
         ),
         (
             (
@@ -48,7 +48,7 @@ def test_depthwise_conv2d_tosa_INT():
                 8,
             ),
             (1, 32, 10, 10),
-            torch.int32,
+            torch.int8,
         ),
     ]
